Leaf-Doctor/requirements.txt
Flask>=3.1.0
gunicorn>=22.0.0
numpy>=1.24.0,<2.3
pandas>=2.0.0,<2.3.0
Pillow>=10.0.0
torch>=2.0.0
torchvision>=0.15.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
sentence-transformers>=3.0.0
groq>=0.9.0
faiss-cpu>=1.7.4
scikit-learn>=1.3.0
seaborn>=0.12.0
matplotlib>=3.7.0
werkzeug>=3.1.0
